MySQL databases backup class - Initial Commit
This is a class that allows you to easily backup a MySQL database (maybe other database types too).
It can easily create daily backups (more or less frequent is easy to setup).
Features include backing up all tables, or a list of them, backing up only structure for specified tables, and how often to store and keep backups. Outputs can be compressed to save space. It works well with backing up multiple databases.
It requires a PDO connection to a database.
It's well documented, with lots of adjustable options, but works well out of the box.